+++ src/main/java/com/team9ookie/dangdo/controller/ReviewController.java
@@ -1,6 +1,6 @@
 package com.team9ookie.dangdo.controller;
 
-import com.team9ookie.dangdo.dto.ReviewRequestDto;
+import com.team9ookie.dangdo.dto.review.ReviewRequestDto;
 import com.team9ookie.dangdo.dto.review.ReviewResponseDto;
 import com.team9ookie.dangdo.service.ReviewService;
 import lombok.RequiredArgsConstructor;
